update on order and checkout conversion raet calculation

+++ B2C_Models/Models/OrderMaster.cs
@@ -23,7 +23,7 @@ namespace B2C_Models.Models
         public int Id { get; set; }
         public string OrderDescription { get; set; }
         public Nullable<int> TotalQuantity { get; set; }
-        public Nullable<int> TotalPrice { get; set; }
+        public Nullable<decimal> TotalPrice { get; set; }
         public int FK_Customer { get; set; }
         public string OrderNo { get; set; }
         public string ShippingAddress { get; set; }
